feat(init): install tiller in cluster
NOTE:

In order for flux and the helm-operator to work successfully, an instance
of tiller must be deployed into the cluster. This is a security vulnerability
in that the tiller account executes with a highly privileged role.

This should be removed once https://github.com/fluxcd/helm-operator/pull/79
has been released. The current version of the helm-operator (1.0.0-rc3) does
not include this PR.